source-service/requirements.txt
fastapi
uvicorn
requests
pydantic
grpcio
grpcio-tools
redis